+++ src/protocol/mod.rs
@@ -5,12 +5,8 @@ pub mod context;
 pub mod ipa;
 pub mod malicious;
 pub mod modulus_conversion;
-#[cfg(feature = "no-prss")]
-mod no_prss;
 pub mod prss;
 pub mod sort;
-#[cfg(not(feature = "no-prss"))]
-mod use_prss;
 
 use crate::error::Error;
 use std::fmt::Debug;
